Fix generation of cipher/decipher URLs
This solves the issue where the cipher/decipher button cannot be used if
Redmine is not deployed on the root URL (see also keeps/redmine_wikicipher#23).

In case of Redmine 2.x, Redmine::Utils.relative_url_root needs to be
prefixed, as it is not a port of the *_path URLs yet.